fix(ui): optimize notification re-renders, isolate mock data, and secure test timers
