+++ src/hooks/useNotificationBell.ts
@@ -14,59 +14,23 @@ export type NotificationItem = {
   unread?: boolean;
 };
 
-export const defaultMockNotifications: NotificationItem[] = [
-  {
-    id: '1',
-    type: 'reservation_new',
-    menteeName: '小明',
-    createdAt: new Date(Date.now() - 1 * 60 * 60 * 1000).toISOString(), // 1 小時前
-    unread: true,
-  },
-  {
-    id: '2',
-    type: 'reservation_success',
-    mentorName: '林導師',
-    createdAt: new Date(Date.now() - 23 * 60 * 60 * 1000).toISOString(), // 23 小時前
-    unread: true,
-  },
-  {
-    id: '3',
-    type: 'reservation_failed',
-    mentorName: '王導師',
-    createdAt: new Date(Date.now() - 24 * 60 * 60 * 1000).toISOString(), // 1 天前
-    unread: false,
-  },
-  {
-    id: '4',
-    type: 'reservation_canceled',
-    mentorName: '陳導師',
-    createdAt: new Date(Date.now() - 5 * 24 * 60 * 60 * 1000).toISOString(), // 5 天前
-    unread: false,
-  },
-  {
-    id: '5',
-    type: 'reservation_upcoming',
-    mentorName: '張導師',
-    createdAt: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000).toISOString(), // 30 天前
-    unread: false,
-  },
-];
-
 export type UseNotificationBellProps = {
   unreadCount: number;
   initialStatus: 'loading' | 'error' | 'empty' | 'success';
   initialNotifications?: NotificationItem[];
+  defaultNotifications?: NotificationItem[];
 };
 
 export function useNotificationBell({
   unreadCount,
   initialStatus,
   initialNotifications,
+  defaultNotifications = [],
 }: UseNotificationBellProps) {
   const [hasBeenClicked, setHasBeenClicked] = React.useState(false);
   const [status, setStatus] = React.useState(initialStatus);
   const [notifications, setNotifications] = React.useState<NotificationItem[]>(
-    () => initialNotifications ?? defaultMockNotifications
+    () => initialNotifications ?? defaultNotifications
   );
 
   const timerRef = React.useRef<NodeJS.Timeout | null>(null);
@@ -94,9 +58,12 @@ export function useNotificationBell({
     if (open) {
       setHasBeenClicked(true);
       setLocalUnreadCount(0);
-      setNotifications((prev) =>
-        prev.map((item) => ({ ...item, unread: false }))
-      );
+      setNotifications((prev) => {
+        const hasUnread = prev.some((item) => item.unread);
+        return hasUnread
+          ? prev.map((item) => ({ ...item, unread: false }))
+          : prev;
+      });
     }
   }, []);
 
@@ -107,10 +74,10 @@ export function useNotificationBell({
     }
     // Simulating a clean reload back to initial or default success list
     timerRef.current = setTimeout(() => {
-      setNotifications(initialNotifications ?? defaultMockNotifications);
+      setNotifications(initialNotifications ?? defaultNotifications);
       setStatus('success');
     }, 1000);
-  }, [initialNotifications]);
+  }, [initialNotifications, defaultNotifications]);
 
   const showBadge = !hasBeenClicked && localUnreadCount > 0;
   const formattedCount =
